chore(handler): refactors ciam package location.
